make category policy,modifine in controller

+++ app/Http/Controllers/CategoryController.php
@@ -6,6 +6,7 @@ use App\Models\Category;
 use App\Http\Requests\StoreCategoryRequest;
 use App\Http\Requests\UpdateCategoryRequest;
 use Auth;
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Support\Str;
 
 class CategoryController extends Controller
@@ -13,8 +14,11 @@ class CategoryController extends Controller
     /**
      * Display a listing of the resource.
      */
-    public function index()
+    public function index(Category $category)
     {
+        if (Gate::denies('view',$category)) {
+            return abort(403);
+        }
         $categories = Category::latest("id")->get();
         return view('category.index',compact('categories'));
     }
@@ -30,8 +34,11 @@ class CategoryController extends Controller
     /**
      * Store a newly created resource in storage.
      */
-    public function store(StoreCategoryRequest $request)
+    public function store(StoreCategoryRequest $request,Category  $category)
     {
+        if (Gate::denies('create',$category)) {
+            return abort(403);
+        }
         $title = $request->title;
         $slug = Str::slug($title);
         $category = new Category();
@@ -45,7 +52,7 @@ class CategoryController extends Controller
     /**
      * Display the specified resource.
      */
-    public function show(Category $category)
+    public function show()
     {
         return redirect()->route('category.index');
     }
@@ -63,6 +70,9 @@ class CategoryController extends Controller
      */
     public function update(UpdateCategoryRequest $request, Category $category)
     {
+        if (Gate::denies('update',$category)) {
+            return abort(403);
+        }
         $title = $request->title;
         $slug = Str::slug($title);
         $category->title = $title;
@@ -77,6 +87,9 @@ class CategoryController extends Controller
      */
     public function destroy(Category $category)
     {
+        if (Gate::denies('delete',$category)) {
+            return abort(403);
+        }
         $category->delete();
         return redirect()->route('category.index')->with('status','Category Successfully Deleted');
     }
